extract auth_witness
