+++ yarn-project/accounts/src/defaults/account_interface.ts
@@ -1,7 +1,8 @@
 import type { AztecAddress, NodeInfo } from '@aztec/aztec.js';
 import { type AccountInterface, type AuthWitnessProvider, CompleteAddress } from '@aztec/aztec.js/account';
 import { type EntrypointInterface, type ExecutionRequestInit } from '@aztec/aztec.js/entrypoint';
-import { type AuthWitness, type TxExecutionRequest } from '@aztec/circuit-types';
+import { type TxExecutionRequest } from '@aztec/circuit-types';
+import { type AuthWitness } from '@aztec/circuit-types/auth-witness';
 import { DefaultAccountEntrypoint } from '@aztec/entrypoints/account';
 import { Fr } from '@aztec/foundation/fields';
