+++ yarn-project/accounts/src/ecdsa/ecdsa_k/account_contract.ts
@@ -1,5 +1,5 @@
 import { type AuthWitnessProvider } from '@aztec/aztec.js/account';
-import { AuthWitness } from '@aztec/circuit-types';
+import { AuthWitness } from '@aztec/circuit-types/auth-witness';
 import { type ContractArtifact } from '@aztec/circuits.js/abi';
 import { CompleteAddress } from '@aztec/circuits.js/contract';
 import { Ecdsa } from '@aztec/foundation/crypto';
