+++ yarn-project/aztec.js/src/entrypoint/entrypoint.ts
@@ -1,10 +1,5 @@
-import {
-  type AuthWitness,
-  type Capsule,
-  type FunctionCall,
-  type HashedValues,
-  type TxExecutionRequest,
-} from '@aztec/circuit-types';
+import { type Capsule, type FunctionCall, type HashedValues, type TxExecutionRequest } from '@aztec/circuit-types';
+import { type AuthWitness } from '@aztec/circuit-types/auth-witness';
 import { type Fr } from '@aztec/foundation/fields';
 
 import { EntrypointPayload, type FeeOptions, computeCombinedPayloadHash } from './payload.js';
