+++ yarn-project/circuit-types/src/auth_witness/auth_witness.ts
