+++ yarn-project/circuit-types/src/auth_witness/index.ts
@@ -1,0 +1,1 @@
+export * from './auth_witness.js';
